FEATURES 5 & 6 COMPLETE - TOTAL NODES/BE PASTEABLE
add two awks to extract the number of total nodes from the temp files.

total nodes extracted by awking for "class=.all" then extracting the
value shown 2 lines later. using the . was a bit of a cheat as i
couldn't work out how to put a ' in regex.

feature 6 was always true