FEATURE 9 COMPLETE - GET LAST REPORT
add 2 functions which both grab the list of pending nodes and awks the
relevant tmp file to lift the last report URL for each node. the
function then goes on to print the node name, curl its last report and
grep all relevant lines from within an awk'ed group of lines. the awk'ed
group is anything between "Pending (" and the next occurrence of "<h3>"

awk - select lines between two patterns

awk '/pattern1/{flag=1 ; next}/pattern2/{flag=0} flag' input_file

this sets the flag to 1 when the pattern is found, and moves to the next
line. when pattern 2 is found, flag is set to 0. the final flag invokes
awk's default behaviour of printing $0 (or the whole line)