ADD CHANGES PER NODE METRIC TO get_stg() get_prd()
rather than simply grepping for the node, grep for the node and print
the number of changes - which is found 9 lines later. use grep to print
only the matched text, and the following 9 lines. use sed to delete the
intermediate 8 lines. use printf to format the first result to 37 chars.